Screens: Navigate: Re-design search Page with Optimized Channel LogoUrl obtention method
Part of why this search page had the design it had was because attaching the logo image
of the channel under the video thumbnail was more complicated than what it seems, the
method used is compute expensive so running a future for each result to obtain the channel
logo image url was hanging out the scroll.

This time I implemented a channel logo list "cache", where each object has the channel name
and the logo Url, upon receiving and showing the search results a future is ran for each of
them to check wheter or not the channel name exist in the cache and if it does, returns the
logo Url.

If the channel name is not cached, the function will spawn an Isolate and send the VideoId
to it, then this Isolate will process the Channel by the Id sent, return only the channel
logo Url and after that check again if the cache has the channel name, if it still does not
then save it.

This way because how the ListView.builder works, instead of running an Isolate every time
an Item becomes rendered it checks and returns the logo Url already cached for that
specific channel.