Update languagePt-BR.dart
I noticed I forgot to add other strings, these are the messages that appear when we click on some option of the 3-dots menu of the videos (e.g. "Video added to Favorites" that appears when clicking to add the video to favorites). Could you add them later? Maybe i'll be missing one or the other but I'll report when I find.