Greatly optimize songs fetch function & implement new PaletteLoader widget
When fetching all devices song, for each a new thumbnail is generated, which takes
about 50ms - 80ms on average, followed by generating a ColorsPalette for that song,
which takes around 300ms - 600ms on average.

Removing the ColorsPalette generation from this process improves the time it takes
to the song fetch function to finish, to a reasonable wait depending on the music
library.

And to resolve the ColorsPalette generation, a new PaletteLoader widget have been
introduced, which loads and saves the palette colors from any given song just when
it's needed.